[Refactor] #95 송신과 수신 역할 분리
