+++ seat-service/src/main/java/com/ticketrush/boundedcontext/seat/app/support/SeatStatusStreamSubscriber.java
@@ -1,0 +1,8 @@
+package com.ticketrush.boundedcontext.seat.app.support;
+
+import org.springframework.web.servlet.mvc.method.annotation.SseEmitter;
+
+public interface SeatStatusStreamSubscriber {
+
+  SseEmitter subscribe(Long performanceId);
+}
